refactor(headers): Unify static/module routes under siteRoute / siteLayout
