+++ src/Layouts/FallbackLayout.tsx
@@ -1,9 +1,11 @@
-type FallbackLayoutProps = {};
+import type { ReactNode } from "react";
 
-export function FallbackLayout({}: FallbackLayoutProps) {
+type FallbackLayoutProps = { children: ReactNode };
+
+export function FallbackLayout({ children }: FallbackLayoutProps) {
   return (
-    <div>
-        
+    <div className="w-dvw h-dvh flex flex-col gap-4 items-center justify-center bg-ludoGrayDark">
+      {children}
     </div>
   );
 }
