+++ src/Layouts/LessonLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "@tanstack/react-router";
-import { LessonHeader } from "../features/Lesson/LessonHeader";
+import { HeaderWithProgress } from "../features/Lesson/HeaderWithProgress";
 import { lessonRoute, lessonSectionRoute } from "../routes/router";
 import { LessonContext } from "../features/Lesson/useLessonContext";
 import { LessonFooter } from "../components/Molecules/Footer/LessonFooter";
@@ -9,7 +9,7 @@ import { useExerciseFlow } from "../Hooks/Logic/Exercises/useExerciseFlow";
 import { ExitDialog } from "@/components/Molecules/Dialog/ExitDialog";
 import { useModal } from "@/Hooks/UI/useModal";
 
-export function LessonSectionLayout() {
+export function LessonLayout() {
   const { exercises, lesson } = lessonSectionRoute.useLoaderData();
   const { exercise: position } = lessonRoute.useSearch();
   const exercisePosition = Number(position ?? 1);
@@ -26,7 +26,7 @@ export function LessonSectionLayout() {
   return (
     <LessonContext.Provider value={state}>
       <MainGridWrapper className="max-h-dvh" gridRows="FULL">
-        <LessonHeader
+        <HeaderWithProgress
           total={exercises.length}
           position={exercisePosition - 1}
           onExit={() => openExit()}
