+++ src/Layouts/SiteLayout.tsx
@@ -1,8 +1,8 @@
-import { Outlet } from "@tanstack/react-router";
-import { NavigationFooter } from "../../components/Molecules/Footer/NavigationFooter";
-import { MainGridWrapper } from "../Grids/MainGridWrapper";
-import { AppHeader } from "../../components/Molecules/Header/AppHeader";
-import { StatsContext } from "../../features/Common/StatsContext";
+import { Outlet, useMatches } from "@tanstack/react-router";
+import { NavigationFooter } from "../components/Molecules/Footer/NavigationFooter";
+import { MainGridWrapper } from "./Grids/MainGridWrapper";
+import { AppHeader } from "../components/Molecules/Header/AppHeader";
+import { StatsContext } from "../features/Common/StatsContext";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { qo } from "@/Hooks/Queries/Definitions/queries";
 
@@ -14,10 +14,15 @@ export function SiteLayout() {
   const { current } = streakPacket;
   const { coins } = coinPacket;
 
+  const matches = useMatches();
+  const active = matches[matches.length - 1];
+  const title =
+    (active?.staticData as { headerTitle?: string })?.headerTitle ?? "LudoCode";
+
   return (
     <StatsContext.Provider value={{ coins: coins, streak: current }}>
       <MainGridWrapper gridRows={"SITE"}>
-        <AppHeader title="Title" />
+        <AppHeader title={title} />
         <Outlet />
         <NavigationFooter />
       </MainGridWrapper>
