+++ src/features/Error/ErrorPage.tsx
@@ -1,5 +1,6 @@
 import { Button } from "@/components/ui/button";
 import { errorMap, type ErrorStatus } from "@/constants/static-data/ErrorData";
+import { FallbackLayout } from "@/Layouts/FallbackLayout";
 
 type ErrorPageProps = { errorCode: ErrorStatus };
 
@@ -8,7 +9,7 @@ export function ErrorPage({ errorCode }: ErrorPageProps) {
     errorMap[errorCode];
 
   return (
-    <div className="w-dvw h-dvh flex flex-col gap-4 items-center justify-center bg-ludoGrayDark">
+    <FallbackLayout>
       <h1 className="text-center text-6xl text-ludoAltText">{status}</h1>
       <div className="p-6 text-center flex flex-col gap-6 rounded-lg bg-ludoGrayLight/70">
         <div className="flex flex-col gap-2">
@@ -18,6 +19,6 @@ export function ErrorPage({ errorCode }: ErrorPageProps) {
 
         <Button onClick={() => fallbackAction?.()}>{actionText}</Button>
       </div>
-    </div>
+    </FallbackLayout>
   );
 }
